[Development] Implemented `ComponentTable` struct
Implemented `ComponentTable` struct which is a column-oriented structure-of-arrays based storage that maps `GameObject`s to their `Component`s.
Interface of this struct is not stable, for example there is only one method for retrieving components. Also, `gamecore::storages` submodule will become private in future, because most of the interface should not be used by the end user and should instead be incapsulated in the `Scene` struct.
Currently, no testing is performed - that is a subject for upcoming commit.