[Development] Reworked `UnpreparedSystem`/`System` API
Reworked `UnpreparedSystem`/`System` API, now there are few structs that represent different types of systems, some of which need to be prepared via `PrepareSystem` trait before becoming `PreparedSystem`. This approach allows to not store local state of the system separately and, instead, inject it into system itself.
Current state of code is very clunky due to multiple trait bounds - going to hide them away with macros (documentation will still be susceptible to this, but that could be tolerated).