[Development] Improved `ggengine::mathcore` module
Improved `ggengine::mathcore` module by revisiting some of the implementations in `mathcore::transforms` and `mathcore::shapes`.
Thought about adding `Polygon` shape to the `mathcore::shapes`, but I think that it is too complicated:
* checks on 0/1-sided polygons are not nice
* need to check for simple polygons, which is computationally expensive
* distinction about convex and concave polygons could be made with a newtype, but that is a lot of boilerplate
* polygon could have custom rotation axis, which makes it very different from `Rect` in that perspective
All in all, implementation of `Polygon` would either be done in future versions, or be delegated to user of a library.